restructuring package for just one folder
freebayes is the only too available in freebayes, so subfolder structure 
is not required
