--- janis_bioinformatics/tools/freebayes/freebayes/freebayes_1_2.py
@@ -1,6 +1,0 @@
-from ..freebayes_1_2 import FreeBayes_1_2
-from .base import FreeBayesBase
-
-
-class FreeBayes_1_2(FreeBayes_1_2, FreeBayesBase):
-    pass
