--- janis_bioinformatics/tools/freebayes/freebayes/latest.py
@@ -1,10 +1,0 @@
-from ..freebayeslatest import FreeBayesLatest
-from .base import FreeBayesBase
-
-
-class FreeBayesLatest(FreeBayesLatest, FreeBayesBase):
-    pass
-
-
-if __name__ == "__main__":
-    print(FreeBayesLatest().help())
