+++ janis_bioinformatics/tools/freebayes/freebayes_1_2.py
@@ -1,4 +1,7 @@
-class FreeBayes_1_2:
+from .base import FreeBayesBase
+
+
+class FreeBayes_1_2(FreeBayesBase):
     @staticmethod
     def container():
         return "papaemmelab/docker-freebayes:latest"
@@ -6,3 +9,6 @@ class FreeBayes_1_2:
     @staticmethod
     def version():
         return "1.2"
+
+    if __name__ == "__main__":
+        print(FreeBayesLatest().help())
